First-APP/requirements.txt
flask==2.3.3
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
gunicorn
waitress
requests
rq
rq-win